+++ client/src/component/pages/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Nav from "../../component/Nav/Nav";
+import Nav from "../../component/Nav";
 
 import indexCSS from './../Index/Index.module.css';
